requirements.txt
Django==1.11.7
django-autoslug==1.9.3
djangorestframework==3.7.3
pytz==2017.3
dj-database-url==0.4.1
gunicorn==19.6.0
whitenoise==3.2
